backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0

# ML — pinned to match premium_model.pkl training environment
scikit-learn==1.7.2
xgboost==2.0.3
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
apscheduler==3.10.4
